Renaming 'configuration' to 'config'
