+++ disk-buffering/src/main/java/io/opentelemetry/contrib/disk/buffering/config/TemporaryFileProvider.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.contrib.disk.buffering.configuration;
+package io.opentelemetry.contrib.disk.buffering.config;
 
 import java.io.File;
 import java.io.IOException;
