+++ disk-buffering/src/test/java/io/opentelemetry/contrib/disk/buffering/internal/storage/TestData.java
@@ -5,8 +5,8 @@
 
 package io.opentelemetry.contrib.disk.buffering.internal.storage;
 
-import io.opentelemetry.contrib.disk.buffering.configuration.StorageConfiguration;
-import io.opentelemetry.contrib.disk.buffering.configuration.TemporaryFileProvider;
+import io.opentelemetry.contrib.disk.buffering.config.StorageConfiguration;
+import io.opentelemetry.contrib.disk.buffering.config.TemporaryFileProvider;
 import io.opentelemetry.contrib.disk.buffering.internal.files.DefaultTemporaryFileProvider;
 import java.io.File;
